backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
faster-whisper
ollama
soundfile
google-generativeai
edge-tts
python-dotenv
